+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Made `dbshell` use `sslcert`, `sslkey`, `sslmode`, and `options` from the
   `'OPTIONS'` part of the `DATABASES` setting.
+- Added support for CockroachDB 22.2, which adds support for
+  `QuerySet.select_for_update(skip_locked=True)`.
 
 ## 4.1 - 2022-08-03